+++ packages/ui/IconButton/IconButton.js
@@ -46,7 +46,7 @@ const IconButtonBase = kind({
 		/**
 		 * The component used to render the [icon]{@link ui/IconButton.IconButtonBase.icon}.
 		 *
-		 * This component will receive the `size` property set on the `IconButton` as well as the
+		 * This component will receive the `flip` and `size` property set on the `IconButton` as well as the
 		 * `icon` class to customize its styling.
 		 *
 		 * @type {Component}
@@ -94,6 +94,14 @@ const IconButtonBase = kind({
 		 * @public
 		 */
 		disabled: PropTypes.bool,
+
+		/**
+		 * Flip the icon horizontally, vertically or both.
+		 *
+		 * @type {('both'|'horizontal'|'vertical')}
+		 * @public
+		 */
+		flip: PropTypes.string,
 
 		/**
 		 * The icon displayed within the IconButton.
@@ -153,7 +161,7 @@ const IconButtonBase = kind({
 		className: ({size, styler}) => styler.append(size)
 	},
 
-	render: ({buttonComponent, children, css, icon, iconComponent: Icon, size, ...rest}) => {
+	render: ({buttonComponent, children, css, flip, icon, iconComponent: Icon, size, ...rest}) => {
 		// To support the simpler use case of only specifying the icon as the children within
 		// <IconButton>, this falls back on using children if icon isn't specified.
 		if (!icon && children) {
@@ -167,7 +175,7 @@ const IconButtonBase = kind({
 			size: size,
 			minWidth: false,
 			children: [
-				<Icon key="icon" size={size} className={css.icon}>{icon}</Icon>,
+				<Icon key="icon" flip={flip} size={size} className={css.icon}>{icon}</Icon>,
 				...React.Children.toArray(children)
 			]
 		});